% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-package.R
\docType{package}
\name{flat-package}
\alias{flat}
\alias{flat-package}
\title{flat: Smooths for mgcv that are constrained to be flat at the ends}
\description{
Smooths for use with mgcv that are flat at one or both of the end knots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/flat}
  \item Report bugs at \url{https://github.com/SWotherspoon/flat/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

Authors:
\itemize{
  \item Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})
  \item Ben Raymond \email{Ben.Raymond@aad.gov.au}
}

}
\keyword{internal}
