% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_package}
\alias{flatten_package}
\title{Produce script reproduce a package}
\usage{
flatten_package(dir = ".", file = "")
}
\arguments{
\item{dir}{source dir, current dir by default.}

\item{file}{destination path, if absent the script is printed to the comnsole.}
}
\value{
Returns \code{NULL} invisibly, called for side effects.
}
\description{
Since the file produce is a self contained flat file you can email it and source
it in the desired folder. {flat} is not needed on the recipient machine.
}
\details{
We ignore :
\itemize{
\item hidden \verb{.Rproj.user*} files
\item git files
\item binary files
\item files under man/ (we assume they can be rebuilt with roxygen)
\item README.md (we assume it can be rebuilt from a README.Rmd)
\item .Rhistory
\item testthat snapshots
}

We can do a bi
}
