% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat.R
\name{Predict.matrix.fb.smooth}
\alias{Predict.matrix.fb.smooth}
\title{GAM Predict Matrix Method}
\usage{
\method{Predict.matrix}{fb.smooth}(object, data)
}
\arguments{
\item{object}{a smooth object, usually generated by a
\code{\link[mgcv:smooth.construct]{mgcv::smooth.construct()}} method having processed a smooth
specification object generated by an \code{\link[mgcv:s]{mgcv::s()}} or
\code{\link[mgcv:te]{mgcv::te()}} term in a \code{\link[mgcv:gam]{mgcv::gam()}} formula.}

\item{data}{A data frame containing the values of the (named)
covariates at which the smooth term is to be evaluated. Exact
requirements are as for \code{\link[mgcv:smooth.construct]{mgcv::smooth.construct()}}}
}
\value{
A matrix mapping the coefficients for the smooth term to its
values at the supplied data values.
}
\description{
A method for \code{\link[mgcv:Predict.matrix]{mgcv::Predict.matrix()}} for penalized cubic
regression splines smooths.
}
\details{
This method constructs a prediction matrix for penalized cubic
regression splines smooths constrained to have zero first
derivative at the leftmost and rightmost knots.
}
