% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat.R
\name{Predict.matrix.zb.smooth}
\alias{Predict.matrix.zb.smooth}
\title{GAM Predict Matrix Method}
\usage{
Predict.matrix.zb.smooth(object, data)
}
\arguments{
\item{object}{a smooth object, usually generated by a
\code{\link{smooth.construct}} method having processed a smooth
specification object generated by an \code{\link{s}} or
\code{\link{te}} term in a \code{\link{gam}} formula.}

\item{data}{A data frame containing the values of the (named)
covariates at which the smooth term is to be evaluated. Exact
requirements are as for \code{\link{smooth.construct}}}
}
\value{
A matrix mapping the coefficients for the smooth term to its
values at the supplied data values.
}
\description{
A method for \code{\link{Predict.matrix}} for penalized cubic
regression splines smooths.
}
\details{
This method constructs a prediction matrix for penalized cubic
regression splines smooths constrained to have zero first
derivative at the leftmost and rightmost knots.
}
