% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat.R
\name{Predict.matrix.zl.smooth}
\alias{Predict.matrix.zl.smooth}
\title{GAM Predict Matrix Method}
\usage{
\method{Predict.matrix}{zl.smooth}(object, data)
}
\arguments{
\item{object}{a smooth object, usually generated by a
\code{\link[=smooth.construct]{smooth.construct()}} method having processed a smooth
specification object generated by an \code{\link[=s]{s()}} or
\code{\link[=te]{te()}} term in a \code{\link[=gam]{gam()}} formula.}

\item{data}{A data frame containing the values of the (named)
covariates at which the smooth term is to be evaluated. Exact
requirements are as for \code{\link[=smooth.construct]{smooth.construct()}}}
}
\value{
A matrix mapping the coefficients for the smooth term to its
values at the supplied data values.
}
\description{
A method for \code{\link[=Predict.matrix]{Predict.matrix()}} for penalized cubic
regression splines smooths.
}
\details{
This method constructs a prediction matrix for penalized cubic
regression splines smooths constrained to have zero first
derivative at the leftmost knot.
}
