% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat.R
\name{smooth.construct.zl.smooth.spec}
\alias{smooth.construct.zl.smooth.spec}
\title{Flat End Penalized Cubic Regression Splines}
\usage{
\method{smooth.construct}{zl.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, usually generated by
a term \code{s(...,bs="fl",...)}}

\item{data}{a list containing just the data (including any by
variable) required by this term, with names corresponding to
\code{object$term} and \code{object$by}. The \code{by} variable
must be the last element.}

\item{knots}{a list containing any knots supplied for basis setup
in same order and with same names as \code{data}. Cannot be
\code{NULL}.}
}
\value{
An object of class "fb.smooth", containing the elements of a
smooth class documented under \code{\link[=smooth.construct]{smooth.construct()}}.
}
\description{
Penalized cubic regression splines constrained to have zero first
derivative at the leftmost knot.
}
\details{
These smooths are similar to the standard "cr" smooths, except
that the first derivative is constrained to be zero at the
leftmost knot. Because the location at which the derivative
constraint is applied is determined by the knots, the knots cannot
be automatically chosen and must be supplied by the user.

Similar to the unconstrained case, the penalty smooth in the
presence of constraints on the first derivative can be shown to be
a cubic spline but with discontinuities in the second derivative
at the points where the derivative constraints apply.  So the
smooth will be a cubic spline within the interval spanned by the
knots, will be linear outside this interval and have
discontinuous second derivatives on the left boundary.
}
