% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{flat-package}
\alias{flat-package}
\title{Flat smooths}
\description{
Provides facilities for estimating broadscale animal motions from
local light level data recorded by an archival tag.
}
\author{
S. Wotherspoon and B. Raymond
}
